requirements.txt
flask
flask-cors
streamlit
requests
pillow
langchain
llama-index
transformers
python-dotenv
